migrated project to gradle-based intellij plugin
