+++ .gitignore
@@ -1,20 +1,8 @@
-# Created by .ignore support plugin (hsz.mobi)
-### Java template
+.idea
+*.iml
 *.class
+.gradle
 
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files #
-*.jar
-*.war
-*.ear
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-
-.idea/
-android-studio-plugin.iml
-.gitignore
-out
-android-studio-plugin.zip
+/target
+/build
+/out
