+++ build.gradle
@@ -1,0 +1,34 @@
+plugins {
+    id 'java'
+    id 'org.jetbrains.intellij' version '0.4.19'
+}
+
+group 'com.crowdin.crowdin-idea'
+version '0.5.10'
+
+sourceCompatibility = 1.8
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+    compile group: 'net.lingala.zip4j', name: 'zip4j', version: '1.3.2'
+    compile files('libs/crowdin-java-sdk-1.0.2.jar')
+}
+
+// See https://github.com/JetBrains/gradle-intellij-plugin/
+intellij {
+    version '2019.2.3'
+    plugins 'git4idea'
+}
+patchPluginXml {
+    changeNotes """
+      Add change notes here.<br>
+      <em>most HTML tags may be used</em>"""
+}
+
+wrapper {
+    gradleVersion = '5.2.1'
+}
