--- resources/application.properties
@@ -1,3 +1,0 @@
-crowdin.base-url=https://api.crowdin.com/api/
-crowdin.project-identifier=
-crowdin.project-key=
