+++ src/main/resources/application.properties
@@ -1,0 +1,3 @@
+crowdin.base-url=https://api.crowdin.com/api/
+crowdin.project-identifier=
+crowdin.project-key=
